requirements.txt
astroid==2.15.1
certifi==2022.12.7
charset-normalizer==3.1.0
dill==0.3.6
idna==3.4
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
platformdirs==3.2.0
pylint==2.17.1
python-dotenv==1.0.0
requests==2.28.2
tomli==2.0.1
tomlkit==0.11.7
typing_extensions==4.5.0
urllib3==1.26.15
wrapt==1.15.0
